+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Services/Taxonomies/AbstractTaxonomy.php
@@ -177,6 +177,6 @@ abstract class AbstractTaxonomy extends AbstractAutomaticallyInstantiatedService
 
     protected function getTaxonomyNamespace(): string
     {
-        return PluginApp::getMainPlugin()->getPluginNamespaceForDB();
+        return PluginApp::getMainPlugin()->getPluginNamespaceForEntityTypeNames();
     }
 }